+++ scripts/build.js
@@ -5,7 +5,7 @@ const { version } = require('../package.json');
 const manifest = require('../chrome-extension/manifest.json');
 
 async function buildExtension() {
-  console.log(`Building venus-devtools version ${version}`);
+  console.log(`Building galacean-devtools version ${version}`);
   let validVersion = version;
   // 处理雨燕 dev 阶段版本号，比如可能长这样 0.0.9001561684-dev.3
   if (version.includes('dev')) {
@@ -15,7 +15,7 @@ async function buildExtension() {
       version.split(sprintId)[0] +
       sprintId.substr(sprintId.length - 3, sprintId.length) +
       deployCount;
-    console.log('Dectected dev version, replace version to', validVersion);
+    console.log('Detected dev version, replace version to', validVersion);
   }
   writeFileSync(
     path.resolve(__dirname, '../chrome-extension/manifest.json'),