+++ src/agent/utils/send-message.js
@@ -1,7 +1,7 @@
 export const sendMessage = function (name, data) {
   window.postMessage(
     {
-      source: 'venus-devtools-agent',
+      source: 'galacean-devtools-agent',
       name: name,
       data: data || {},
     },