+++ zip.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-mkdir venus-devtools
-cp -R chrome-extension/* venus-devtools/
-zip -r venus-devtools-$(date +%m%d%H%M).zip venus-devtools/
-rm -rf venus-devtools/
+mkdir galacean-devtools
+cp -R chrome-extension/* galacean-devtools/
+zip -r galacean-devtools-$(date +%m%d%H%M).zip galacean-devtools/
+rm -rf galacean-devtools/