+++ History.md
@@ -2,5 +2,5 @@
 
 - 展示 Entity 树
 - 展示被选中的 Entity 的属性
-- 被选中的 Entity 可以通过 Console 中的 `$venus` 访问
+- 被选中的 Entity 可以通过 Console 中的 `$galacean` 访问